pst-analysis-engine/worker/requirements.txt
boto3==1.35.20
opensearch-py==2.6.0
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0
psycopg2-binary==2.9.9
requests==2.32.3
ocrmypdf==16.0.4
pytesseract==0.3.13
Pillow==10.4.0
SQLAlchemy==2.0.35
pydantic-settings==2.6.1
pydantic==2.10.3
